pyproject.toml
[tool.poetry]
name = "arxiv-telegram"
version = "0.1.0"
description = ""
authors = ["VladimirMynka <vladimirmynka@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
arxiv = "^2.1.3"
aiosqlite = "^0.20.0"
openai = "^1.54.4"
python-telegram-bot = "20.3"
pypdf2 = "^3.0.1"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
